refactor ModelChangePropagator
